+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.1.0
+
+- Also accept the Request object as first argument of the constructor.
+- Add toJSON method to the instance for easy serialization.
+
 ## 1.0.0
 
 Stable release: No breaking changes